requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
cvzone==1.6.1
mediapipe==0.10.7
firebase-admin==6.2.0
google-generativeai==0.3.2
python-dotenv==1.0.0
Werkzeug==2.3.7
numpy==1.24.3
Pillow==10.0.1 